requirements.txt
torch>=2.3.0
torchvision>=0.18.0
matplotlib
seaborn
scikit-learn
Pillow
numpy
